+++ packages/create-sitecore-jss/src/templates/nextjs-styleguide/src/components/graphql/GraphQL-ConnectedDemo.dynamic.tsx
@@ -7,8 +7,10 @@ import {
   constants,
   GraphQLRequestClient,
   withDatasourceCheck,
-  resetEditorChromes,
 } from '@sitecore-jss/sitecore-jss-nextjs';
+import {
+  resetEditorChromes
+} from '@sitecore-jss/sitecore-jss-nextjs/utils';
 import NextLink from 'next/link';
 import { ConnectedDemoQueryDocument } from './GraphQL-ConnectedDemo.dynamic.graphql';
 import {