+++ packages/create-sitecore-jss/src/templates/nextjs/src/Navigation.tsx
@@ -1,4 +1,4 @@
-import { getPublicUrl } from '@sitecore-jss/sitecore-jss-nextjs';
+import { getPublicUrl } from '@sitecore-jss/sitecore-jss-nextjs/utils';
 
 // Prefix public assets with a public URL to enable compatibility with Sitecore editors.
 // If you're not supporting Sitecore editors, you can remove this.