+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/site-resolver/index.ts
@@ -1,4 +1,4 @@
-import { SiteInfo, SiteResolver } from '@sitecore-jss/sitecore-jss-nextjs/middleware';
+import { SiteInfo, SiteResolver } from '@sitecore-jss/sitecore-jss-nextjs/site';
 import * as plugins from 'temp/site-resolver-plugins';
 
 /*